p2p/discover/v5wire: move low-level crypto to its own file
